requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
blinker==1.9.0
bcrypt==5.0.0
certifi==2026.4.22
cffi==2.0.0
click==8.3.2
colorama==0.4.6
cryptography==47.0.0
dnspython==2.8.0
email-validator==2.3.0
ecdsa==0.19.2
fastapi==0.136.0
Flask==3.1.3
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
passlib==1.7.4
pyasn1==0.6.3
pycparser==3.0
pydantic==2.13.2
pydantic_core==2.46.2
pymongo==4.16.0
python-dotenv==1.2.2
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
Werkzeug==3.1.8
